+++ inst/examples/example.R
@@ -15,7 +15,7 @@ check(res_data, occ_data)
 # Create a `samc-class` object with the resistance and absorption data using
 # the samc() function. We use the recipricol of the arithmetic mean for
 # calculating the transition matrix. Note, the input data here are matrices,
-# not RasterLayers. If using RasterLayers, the latlon parameter must be set.
+# not RasterLayers.
 samc_obj <- samc(res_data, abs_data, tr_fun = function(x) 1/mean(x))